Move form components to separate folder
