+++ packages/frontend/src/components/form/profile_picture.tsx
